+++ addons/mail/static/src/core_ui/message.js
@@ -3,7 +3,7 @@
 import { ImStatus } from "@mail/discuss_app/im_status";
 import { AttachmentList } from "@mail/attachments/attachment_list";
 import { MessageInReply } from "./message_in_reply";
-import { isEventHandled, markEventHandled } from "@mail/utils/misc";
+import { isEventHandled, markEventHandled } from "@web/core/utils/misc";
 import { convertBrToLineBreak, htmlToTextContentInline } from "@mail/utils/format";
 import { MessageReactionMenu } from "@mail/core_ui/message_reaction_menu";
 import {
@@ -23,7 +23,7 @@ import { MessageConfirmDialog } from "./message_confirm_dialog";
 import { LinkPreviewList } from "./link_preview/link_preview_list";
 import { RelativeTime } from "./relative_time";
 import { MessageReactions } from "./message_reactions";
-import { useEmojiPicker } from "../emoji_picker/emoji_picker";
+import { useEmojiPicker } from "@web/core/emoji_picker/emoji_picker";
 import { usePopover } from "@web/core/popover/popover_hook";
 import { MessageNotificationPopover } from "./message_notification_popover";
 import { MessageSeenIndicator } from "./message_seen_indicator";
