+++ addons/mail/static/src/discuss_app/channel_selector.js
@@ -7,7 +7,8 @@ import { useService } from "@web/core/utils/hooks";
 import { getActiveHotkey } from "@web/core/hotkeys/hotkey_service";
 import { Component, onMounted, useRef, useState } from "@odoo/owl";
 import { cleanTerm } from "@mail/utils/format";
-import { createLocalId, isEventHandled, markEventHandled } from "@mail/utils/misc";
+import { createLocalId } from "@mail/utils/misc";
+import { isEventHandled, markEventHandled } from "@web/core/utils/misc";
 import { _t } from "@web/core/l10n/translation";
 
 export class ChannelSelector extends Component {
