+++ addons/mail/static/src/utils/misc.js
@@ -39,37 +39,6 @@ export function isDragSourceExternalFile(dataTransfer) {
     return false;
 }
 
-const eventHandledWeakMap = new WeakMap();
-
-/**
- * Returns whether the given event has been handled with the given markName.
- *
- * @param {Event} ev
- * @param {string} markName
- * @returns {boolean}
- */
-export function isEventHandled(ev, markName) {
-    if (!eventHandledWeakMap.get(ev)) {
-        return false;
-    }
-    return eventHandledWeakMap.get(ev).includes(markName);
-}
-
-/**
- * Marks the given event as handled by the given markName. Useful to allow
- * handlers in the propagation chain to make a decision based on what has
- * already been done.
- *
- * @param {Event} ev
- * @param {string} markName
- */
-export function markEventHandled(ev, markName) {
-    if (!eventHandledWeakMap.get(ev)) {
-        eventHandledWeakMap.set(ev, []);
-    }
-    eventHandledWeakMap.get(ev).push(markName);
-}
-
 /**
  * @param {Object} target
  * @param {string} key
