+++ addons/mail/static/src/views/fields/emojis_field_common/emojis_field_common.js
@@ -1,6 +1,6 @@
 /** @odoo-module **/
 
-import { useEmojiPicker } from "@mail/emoji_picker/emoji_picker";
+import { useEmojiPicker } from "@web/core/emoji_picker/emoji_picker";
 import { useRef } from "@odoo/owl";
 
 /*
