+++ addons/mail/static/tests/helpers/test_utils.js
@@ -2,7 +2,7 @@
 
 import { getPyEnv, startServer } from "@bus/../tests/helpers/mock_python_environment";
 
-import { loadEmoji } from "@mail/emoji_picker/emoji_picker";
+import { loadEmoji } from "@web/core/emoji_picker/emoji_picker";
 import { getAdvanceTime } from "@mail/../tests/helpers/time_control";
 import { patchBrowserNotification } from "@mail/../tests/helpers/patch_notifications";
 import { getWebClientReady } from "@mail/../tests/helpers/webclient_setup";
