+++ addons/web/static/src/core/emoji_picker/emoji_data.js
