+++ addons/web/static/src/core/utils/misc.js
@@ -1,0 +1,31 @@
+/* @odoo-module */
+const eventHandledWeakMap = new WeakMap();
+
+/**
+ * Returns whether the given event has been handled with the given markName.
+ *
+ * @param {Event} ev
+ * @param {string} markName
+ * @returns {boolean}
+ */
+export function isEventHandled(ev, markName) {
+    if (!eventHandledWeakMap.get(ev)) {
+        return false;
+    }
+    return eventHandledWeakMap.get(ev).includes(markName);
+}
+
+/**
+ * Marks the given event as handled by the given markName. Useful to allow
+ * handlers in the propagation chain to make a decision based on what has
+ * already been done.
+ *
+ * @param {Event} ev
+ * @param {string} markName
+ */
+export function markEventHandled(ev, markName) {
+    if (!eventHandledWeakMap.get(ev)) {
+        eventHandledWeakMap.set(ev, []);
+    }
+    eventHandledWeakMap.get(ev).push(markName);
+}
